Refactor: Vueコンポーネント リファクタリング
コンポーネント精査、mapActionsの使用
