--- components/atoms/RestartButton.vue
@@ -1,13 +1,0 @@
-<template>
-  <button @click="onClick">やり直す</button>
-</template>
-
-<script>
-export default {
-  methods: {
-    onClick() {
-      location.reload();
-    }
-  }
-}
-</script>
